move to rebalance_patches, update hashes and add comments
