+++ src/open_prime_rando/echoes/specific_area_patches/rebalance_patches.py
@@ -4,9 +4,15 @@ import logging
 from typing import TYPE_CHECKING
 
 from retro_data_structures.enums.echoes import Message, State
+from retro_data_structures.properties.echoes.archetypes.EditorProperties import EditorProperties
 from retro_data_structures.properties.echoes.archetypes.LayerSwitch import LayerSwitch
+from retro_data_structures.properties.echoes.archetypes.Transform import Transform
+from retro_data_structures.properties.echoes.archetypes.TriggerInfo import FlagsTrigger
+from retro_data_structures.properties.echoes.core.Vector import Vector
 from retro_data_structures.properties.echoes.objects import (
     Actor,
+    CameraHint,
+    CameraShaker,
     IngSpiderballGuardian,
     MemoryRelay,
     Pickup,
@@ -14,6 +20,8 @@ from retro_data_structures.properties.echoes.objects import (
     SpawnPoint,
     Splinter,
     Timer,
+    Trigger,
+    TriggerOrientated,
 )
 
 from open_prime_rando.area_patcher import AreaPatcher, decorate_patcher
@@ -58,6 +66,7 @@ def register_all(area_patcher: AreaPatcher) -> None:
         torvus_temple,
         gfmc_compound,
         sanctuary_entrance,
+        minigyro_terminal_fall,
     ]:
         area_patcher.add_function(func)
 
@@ -267,3 +276,209 @@ def sanctuary_entrance(editor: PatcherEditor, mlvl: Mlvl, area: Area) -> None:
     """
     area.move_instance("Dead Luminoth 4 KeyBearer", "Default")
     area.move_instance("Luminoth Light Support", "Default")
+
+
+@decorate_patcher(SANCTUARY_FORTRESS_MLVL, sanctuary_fortress.MINIGYRO_CHAMBER_MREA)
+def minigyro_terminal_fall(editor: PatcherEditor, mlvl: Mlvl, area: Area) -> None:
+    """
+    Adds a Terminal Fall in Minigyro Chamber to prevent oblivious players from accidentally
+    going out of bounds if traversing through opposite side or have Cannon Ball on the
+    intended side (The out of bounds is still possible if the player unmorphs early, and
+    there is a failsafe reposition in case the oblivious player also happens to unmorph early.)
+    """
+    # Copying objects from Unseen Way
+    other_area = editor.get_area(SANCTUARY_FORTRESS_MLVL, sanctuary_fortress.UNSEEN_WAY_MREA)
+
+    # And placing them in Minigyro Chamber
+    filter_blur = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Camera Blur").get_properties()
+    )
+
+    filter_flash = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance(0x170021).get_properties()
+    )
+
+    camera_hint_frozen = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Frozen").get_properties()
+    )
+
+    camera_hint_morph_ball = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Morph Fall").get_properties()
+    )
+
+    camera_hint_prevent_ledge_avoidance = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Prevent Ledge Avoidance").get_properties()
+    )
+
+    camera_shaker = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Camera Shaker_HArsh_Short").get_properties()
+    )
+    with camera_shaker.edit_properties(CameraShaker) as shaker:
+        shaker.editor_properties.transform.position = Vector(170.0, 182.0, -114.0)
+
+    no_morph_control_hint = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Disable Morph").get_properties()
+    )
+
+    radial_damage = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Hurt Player").get_properties()
+    )
+
+    originator_relay = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Set Player Originator for Hurt").get_properties()
+    )
+
+    flash_controls_relay = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance(0x17001D).get_properties()
+    )
+
+    rumble_effect = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Fall Rumble").get_properties()
+    )
+
+    sequence_timer = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Fall to Death").get_properties()
+    )
+
+    sound_into_hud = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Sound - Into Hud").get_properties()
+    )
+
+    east_spawnpoint = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Fall to Death Return").get_properties()
+    )
+    with east_spawnpoint.edit_properties(SpawnPoint) as east_spawn:
+        east_spawn.editor_properties.transform.position = Vector(150.665604, 130.681992, -116.36132)
+        east_spawn.morphed = True
+
+    west_spawnpoint = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance(0x170022).get_properties()
+    )
+    with west_spawnpoint.edit_properties(SpawnPoint) as west_spawn:
+        west_spawn.editor_properties.transform.position = Vector(150.683228, 161.135986, -116.07415)
+        west_spawn.morphed = True
+
+    player_in_area_special_function = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Death Fall Player In Area").get_properties()
+    )
+
+    flash_timer = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Timer White Flash Duration").get_properties()
+    )
+
+    west_spawnpoint_trigger = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance(0x170004).get_properties()
+    )
+    with west_spawnpoint_trigger.edit_properties(Trigger) as west_spawn_trigger:
+        west_spawn_trigger.editor_properties.transform.position = Vector(120.773407, 149.127029, -106.985382)
+        west_spawn_trigger.editor_properties.transform.scale = Vector(10.0, 10.0, 10.0)
+
+    east_spawnpoint_trigger = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Fall Trigger_Platform").get_properties()
+    )
+    with east_spawnpoint_trigger.edit_properties(TriggerOrientated) as east_spawn_trigger:
+        east_spawn_trigger.editor_properties.transform.position = Vector(182.771896, 142.353943, -107.949036)
+        east_spawn_trigger.editor_properties.transform.scale = Vector(10.0, 10.0, 10.0)
+
+    fall_Trigger = area.get_layer("Gyroscope puzzle").add_instance_with(
+        other_area.get_instance("Fall Trigger").get_properties()
+    )
+    with fall_Trigger.edit_properties(TriggerOrientated) as falling_trigger:
+        falling_trigger.editor_properties.transform.position = Vector(151.665604, 144.968307, -124.108681)
+        falling_trigger.editor_properties.transform.scale = Vector(50.0, 50.0, 3.0)
+        falling_trigger.trigger.flags_trigger &= ~FlagsTrigger.DetectUnmorphedPlayer
+
+    # Additional Extra Objects
+    camera_teleport = area.get_layer("Gyroscope puzzle").add_instance_with(
+        CameraHint(
+            editor_properties=EditorProperties(
+                name="Teleport Player Camera",
+                transform=Transform(
+                    position=Vector(151.658524, 145.002243, -120.17498),
+                    scale=Vector(2.0, 2.0, 2.0),
+                ),
+            ),
+            priority=1,
+            timer=0.02,
+            flags_camera_hint=262462,
+        )
+    )
+    east_side_camerahint_reset_trigger = area.get_layer("Gyroscope puzzle").add_instance_with(
+        Trigger(
+            editor_properties=EditorProperties(
+                name="Onto gyro platforms (tunnel) 2nd Trigger",
+                transform=Transform(
+                    position=Vector(150.885666, 130.688904, -116.225525),
+                    scale=Vector(2.983, 1.44, 5.068),
+                ),
+            )
+        )
+    )
+    reset_camerahints_trigger = area.get_layer("Gyroscope puzzle").add_instance_with(
+        Trigger(
+            editor_properties=EditorProperties(
+                name="Reset CameraHints",
+                transform=Transform(
+                    position=Vector(151.665604, 144.968307, -124.108681),
+                    scale=Vector(50.0, 50.0, 3.0),
+                ),
+            )
+        )
+    )
+    fall_failsafe_trigger = area.get_layer("Gyroscope puzzle").add_instance_with(
+        Trigger(
+            editor_properties=EditorProperties(
+                name="Catch Player Failsafe",
+                transform=Transform(position=Vector(151.665604, 144.968307, -300), scale=Vector(1000.0, 1000.0, 3.0)),
+            )
+        )
+    )
+
+    # Define some existing objects
+    gyro_camera_hint = area.get_instance("Gyro Outer Ring")
+    tunnels_camera_hint = area.get_instance("Tunnels")
+    supertrigger = area.get_instance("Supertrigger enter tunnel")
+
+    # Adding connections from terminal fall related objects that Retro added on every room
+    originator_relay.add_connection(State.Zero, Message.Action, radial_damage)
+    flash_controls_relay.add_connection(State.Zero, Message.ResetAndStart, flash_timer)
+    flash_controls_relay.add_connection(State.Zero, Message.Increment, filter_blur)
+    flash_controls_relay.add_connection(State.Zero, Message.Increment, filter_flash)
+    for target, message in [
+        (flash_controls_relay, Message.SetToZero),
+        (sound_into_hud, Message.Play),
+        (no_morph_control_hint, Message.Decrement),
+        (no_morph_control_hint, Message.Increment),
+        (west_spawnpoint, Message.SetToZero),
+        (camera_hint_prevent_ledge_avoidance, Message.Decrement),
+        (camera_hint_morph_ball, Message.Increment),
+        (camera_hint_frozen, Message.Increment),
+        (originator_relay, Message.SetToZero),
+        (camera_hint_frozen, Message.Decrement),
+        (rumble_effect, Message.Action),
+        (east_spawnpoint, Message.SetToZero),
+        (camera_shaker, Message.Action),
+        (fall_Trigger, Message.Activate),
+    ]:
+        sequence_timer.add_connection(State.Sequence, message, target)
+    player_in_area_special_function.add_connection(State.Entered, Message.Activate, fall_Trigger)
+    player_in_area_special_function.add_connection(State.Exited, Message.Deactivate, fall_Trigger)
+    player_in_area_special_function.add_connection(State.Entered, Message.Activate, fall_failsafe_trigger)
+    player_in_area_special_function.add_connection(State.Exited, Message.Deactivate, fall_failsafe_trigger)
+    flash_timer.add_connection(State.Zero, Message.Decrement, filter_blur)
+    flash_timer.add_connection(State.Zero, Message.Decrement, filter_flash)
+    west_spawnpoint_trigger.add_connection(State.Entered, Message.Deactivate, east_spawnpoint)
+    west_spawnpoint_trigger.add_connection(State.Entered, Message.Activate, west_spawnpoint)
+    fall_Trigger.add_connection(State.Entered, Message.SetOriginator, originator_relay)
+    fall_Trigger.add_connection(State.Entered, Message.Start, sequence_timer)
+    east_spawnpoint_trigger.add_connection(State.Entered, Message.Increment, camera_hint_prevent_ledge_avoidance)
+    east_spawnpoint_trigger.add_connection(State.Exited, Message.Decrement, camera_hint_prevent_ledge_avoidance)
+    east_spawnpoint_trigger.add_connection(State.Entered, Message.Activate, east_spawnpoint)
+    east_spawnpoint_trigger.add_connection(State.Entered, Message.Deactivate, west_spawnpoint)
+    # CameraHint related connections
+    reset_camerahints_trigger.add_connection(State.Entered, Message.Decrement, gyro_camera_hint)
+    reset_camerahints_trigger.add_connection(State.Entered, Message.Decrement, tunnels_camera_hint)
+    reset_camerahints_trigger.add_connection(State.Entered, Message.Increment, camera_teleport)
+    east_side_camerahint_reset_trigger.add_connection(State.Connect, Message.Attach, supertrigger)
+    # Attach failsafe Trigger to main Fall Trigger
+    fall_failsafe_trigger.add_connection(State.Connect, Message.Attach, fall_Trigger)
