+++ tests/test_files/ntsc_hashes.json
@@ -26506,7 +26506,7 @@
             "MREA_7b94b06b": "1561b8ff545f188a3babf8ab3900f28c763a063f9a55046c0b1088d5e855eafd",
             "MREA_844a690c": "024d2765f5ff4526f262b78586ff0def371dc4ff46861538761b55365dce735e",
             "MREA_914f1381": "5abe6c5d394a0cc96cacf321780cc4c967dbd77d0d02845ae4d91a319c004fdc",
-            "MREA_96f4ca1e": "15619cbabc963e393588c9fd7b960bddd95ff6bb261a7af7eb5bac932ad3816b",
+            "MREA_96f4ca1e": "25335e232dd9751c9972a944765534eb9b4fa0a376bd7fb68d47d7714660f9ae",
             "MREA_9d804955": "2e8f64d41613ccd953d27b05acf89946d3a654527b94fde66bb323c33391aa80",
             "MREA_a2406387": "a3e714474462293affcc726174701e0b08708c33cd39e68a4fc9e30f539a9c9c",
             "MREA_a3656292": "a903bb3d9f39147e79d6d53cbabe33086f5349c4289b7f0c2e1ea4de43272484",
