+++ tests/test_files/pal_hashes.json
@@ -26526,7 +26526,7 @@
             "MREA_7b94b06b": "1561b8ff545f188a3babf8ab3900f28c763a063f9a55046c0b1088d5e855eafd",
             "MREA_844a690c": "69d734487518a2d4f12bb95bccf23c93d87f70e91e38ee61b4b05a6addcf6340",
             "MREA_914f1381": "9aeb1f87cb00e2cc1ea14465e70293a1b5f9a67866863fd7c4ada68644537034",
-            "MREA_96f4ca1e": "8022ff3363682a4695872da2a6be49d1769bd4ab17fe0cba5570d1742839bfbd",
+            "MREA_96f4ca1e": "daa5466090c54c111f954385f0dcd69f529ac2aa5f671d8945581c183878c9c2",
             "MREA_9d804955": "8b761d6d68ffb57dc512512322c09b015c26b337e5c8892cbe2789c3142d8b2b",
             "MREA_a2406387": "262eb6c729a25f5842d740898149ef3684a00d4212412d3dc97e56621c5f71f1",
             "MREA_a3656292": "a3f72d82e4b0bbc778b07b265ef4e117465f921ef1339c7287585ca56dc74e04",
